move logic to user session
